+++ source/Public/Config/Set-FabricConfig.ps1
@@ -1,10 +1,10 @@
 function Set-FabricConfig {
     <#
 .SYNOPSIS
-Register the configuration for use with all functions in the PSFabricTools module.
+Register the configuration for use with all functions in the FabricTools module.
 
 .DESCRIPTION
-Register the configuration for use with all functions in the PSFabricTools module.
+Register the configuration for use with all functions in the FabricTools module.
 
 .PARAMETER WorkspaceGUID
 This is the workspace GUID in which the Data Warehouse resides.
@@ -19,14 +19,14 @@ Defaults to api.powerbi.com
 If set, the configuration will not be persisted to the registry.
 
 .EXAMPLE
-    Registers the specified Fabric Data Warehouse configuration for use with all functions in the PSFabricTools module.
+    Registers the specified Fabric Data Warehouse configuration for use with all functions in the FabricTools module.
 
     ```powershell
     Set-FabricConfig -WorkspaceGUID 'GUID-GUID-GUID-GUID' -DataWarehouseGUID 'GUID-GUID-GUID-GUID'
     ```
 
 .EXAMPLE
-    Registers the specified Fabric Data Warehouse configuration for use with all functions in the PSFabricTools module, but does not persist the values.
+    Registers the specified Fabric Data Warehouse configuration for use with all functions in the FabricTools module, but does not persist the values.
 
     ```powershell
     Set-FabricConfig -WorkspaceGUID 'GUID-GUID-GUID-GUID' -DataWarehouseGUID 'GUID-GUID-GUID-GUID' -SkipPersist
@@ -50,18 +50,18 @@ Author: Jess Pomfret
     if ($PSCmdlet.ShouldProcess("Setting Fabric Configuration")) {
 
         if ($BaseUrl) {
-            Set-PSFConfig -Module PSFabricTools -Name BaseUrl -Value $BaseUrl
+            Set-PSFConfig -Module 'FabricTools' -Name BaseUrl -Value $BaseUrl
         }
         if ($WorkspaceGUID) {
-            Set-PSFConfig -Module PSFabricTools -Name WorkspaceGUID -Value $WorkspaceGUID
+            Set-PSFConfig -Module 'FabricTools' -Name WorkspaceGUID -Value $WorkspaceGUID
         }
         if ($DataWarehouseGUID) {
-            Set-PSFConfig -Module PSFabricTools -Name DataWarehouseGUID -Value $DataWarehouseGUID
+            Set-PSFConfig -Module 'FabricTools' -Name DataWarehouseGUID -Value $DataWarehouseGUID
         }
 
         # Register the config values in the registry if skip persist is not set
         if (-not $SkipPersist) {
-            Register-PSFConfig -Module PSFabricTools -Scope SystemMandatory
+            Register-PSFConfig -Module 'FabricTools' -Scope SystemMandatory
         }
     }
 }